+++ android/app/src/main/kotlin/com/victorkariuki/resistorcalculator/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.victorkariuki.resistorcalculator
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()
+